rider reformat code

+++ Hedra/Engine/Management/DataManager.cs
@@ -6,46 +6,84 @@
  * 
  * To change this template use Tools | Options | Coding | Edit Standard Headers.
  */
+
 using System;
-using System.Numerics;
-using System.IO;
-using Hedra.Engine.Player;
-using Hedra.Engine.Generation;
 using System.Collections.Generic;
-using System.Linq;
+using System.IO;
+using System.Numerics;
 using Hedra.Engine.ClassSystem;
 using Hedra.Engine.IO;
 using Hedra.Engine.ItemSystem;
+using Hedra.Engine.Player;
 using Hedra.Engine.QuestSystem;
-using Hedra.Engine.Rendering.UI;
 using Hedra.Items;
 using Hedra.Mission;
 
 namespace Hedra.Engine.Management
 {
     /// <summary>
-    /// Description of SaveManager.
+    ///     Description of SaveManager.
     /// </summary>
     public static class DataManager
     {
-        public static string CharactersFolder => $"{AssetManager.AppPath}/Characters/";
         private const float SaveVersion = 1.61f;
-        
+        public static string CharactersFolder => $"{AssetManager.AppPath}/Characters/";
+
+        public static PlayerInformation[] PlayerFiles
+        {
+            get
+            {
+                var filesList = new List<PlayerInformation>();
+                var files = Directory.GetFiles(CharactersFolder);
+
+                for (var i = 0; i < files.Length; i++)
+                {
+                    if (files[i].EndsWith(".bak"))
+                        continue;
+                    try
+                    {
+                        var information = LoadPlayer(files[i]);
+                        if (information == null)
+                        {
+                            File.Delete(files[i]);
+                            if (File.Exists(files[i] + ".bak")) File.Delete(files[i] + ".bak");
+                            continue;
+                        }
+
+                        filesList.Add(information);
+                    }
+                    catch (Exception e)
+                    {
+                        if (e is UnauthorizedAccessException)
+                        {
+                            Log.WriteLine("Unauthorized access in file: " + Path.GetFileName(files[i]));
+                            continue;
+                        }
+
+                        Log.WriteLine(e.ToString());
+                    }
+                }
+
+                return filesList.ToArray();
+            }
+        }
+
+
+        public static int CharacterCount => PlayerFiles.Length;
+
         public static void SavePlayer(PlayerInformation Information)
         {
             var chrFile = $"{CharactersFolder}{Information.Name}";
-            
-            if(File.Exists(chrFile + ".db"))
+
+            if (File.Exists(chrFile + ".db"))
             {
-                if (File.Exists(chrFile + ".db.bak"))
-                {
-                    File.Delete(chrFile + ".db.bak");
-                }
+                if (File.Exists(chrFile + ".db.bak")) File.Delete(chrFile + ".db.bak");
                 File.Copy(chrFile + ".db", chrFile + ".db.bak");
-                
+
                 var fInfo = new FileInfo(chrFile + ".db.bak");
                 fInfo.Attributes |= FileAttributes.Hidden;
             }
+
             using (var fs = File.Create(CharactersFolder + Information.Name + ".db"))
             {
                 using (var bw = new BinaryWriter(fs))
@@ -67,7 +105,7 @@ namespace Hedra.Engine.Management
 
                     bw.Write(Information.SkillsData.Length);
                     bw.Write(Information.SkillsData);
-                    
+
                     bw.Write(Information.RealmData.Length);
                     bw.Write(Information.RealmData);
 
@@ -89,30 +127,29 @@ namespace Hedra.Engine.Management
                             bw.Write(itemBytes);
                         }
                     }
+
                     var recipes = Information.Recipes;
                     if (recipes != null)
                     {
                         bw.Write(recipes.Length);
-                        for(var i = 0; i < recipes.Length; ++i)
-                        {
-                            bw.Write(recipes[i]);
-                        } 
+                        for (var i = 0; i < recipes.Length; ++i) bw.Write(recipes[i]);
                     }
+
                     var quests = Information.Quests;
                     if (quests != null)
                     {
                         bw.Write(quests.Length);
-                        for(var i = 0; i < quests.Length; ++i)
+                        for (var i = 0; i < quests.Length; ++i)
                         {
                             var questBytes = quests[i].ToArray();
                             bw.Write(questBytes.Length);
                             bw.Write(questBytes);
-                        } 
+                        }
                     }
                 }
             }
         }
-    
+
         public static PlayerInformation DataFromPlayer(IPlayer Player)
         {
             var data = new PlayerInformation
@@ -136,17 +173,17 @@ namespace Hedra.Engine.Management
 
             return data;
         }
-        
+
         public static PlayerInformation LoadPlayer(string FileName)
         {
             var info = new FileInfo(FileName);
             FileName = info.Length == 0 ? $"{FileName}.bak" : FileName;
-            var data = DataManager.LoadPlayer(File.Open(FileName, FileMode.Open));
+            var data = LoadPlayer(File.Open(FileName, FileMode.Open));
             if (data == null)
             {
                 Log.WriteLine($"Failed to load character {FileName}. It might be incompatible.");
             }
-            else if(data.IsCorrupt)
+            else if (data.IsCorrupt)
             {
                 Log.WriteLine($"[IO] Detected corrupt character file '{Path.GetFileName(FileName)}'.");
                 if (!File.Exists($"{FileName}.bak"))
@@ -154,6 +191,7 @@ namespace Hedra.Engine.Management
                     Log.WriteLine($"[IO] Failed to load character '{Path.GetFileName(FileName)}'.");
                     return null;
                 }
+
                 Log.WriteLine("[IO] Character backup found! Retrying...");
                 File.Delete(FileName);
                 File.Copy($"{FileName}.bak", FileName);
@@ -162,7 +200,8 @@ namespace Hedra.Engine.Management
                     Attributes = FileAttributes.Normal
                 };
                 return LoadPlayer($"{FileName}.bak");
-            }        
+            }
+
             return data;
         }
 
@@ -172,27 +211,23 @@ namespace Hedra.Engine.Management
             var useDefaultCustomization = false;
             using (var br = new BinaryReader(Str))
             {
-                float version = br.ReadSingle();
+                var version = br.ReadSingle();
                 if (version < 1.6f) return null;
                 information.Name = br.ReadString();
                 if (version >= 1.61f)
-                {
                     information.Customization = CustomizationData.Read(br);
-                }
                 else
-                {
                     useDefaultCustomization = true;
-                }
                 information.Rotation = new Vector3(br.ReadSingle(), br.ReadSingle(), br.ReadSingle());
                 information.Health = br.ReadSingle();
                 information.Xp = br.ReadSingle();
                 information.Level = br.ReadInt32();
                 information.Mana = br.ReadSingle();
-                
+
                 information.SkillsData = br.ReadBytes(br.ReadInt32());
-                information.RealmData = br.ReadBytes(br.ReadInt32()); 
+                information.RealmData = br.ReadBytes(br.ReadInt32());
                 information.ToolbarData = br.ReadBytes(br.ReadInt32());
-                
+
                 information.Class = ClassDesign.FromString(br.ReadString());
                 information.RandomFactor = br.ReadSingle();
 
@@ -200,9 +235,10 @@ namespace Hedra.Engine.Management
                 information.Recipes = LoadRecipes(br);
                 information.Quests = LoadQuests(br);
             }
+
             Str.Close();
             Str.Dispose();
-            if(useDefaultCustomization)
+            if (useDefaultCustomization)
                 information.Customization = CustomizationData.FromClass(information.Class, HumanGender.Male);
             return information;
         }
@@ -211,25 +247,26 @@ namespace Hedra.Engine.Management
         {
             var quests = new List<SerializedQuest>();
             var length = Reader.ReadInt32();
-            for(var i = 0; i < length; ++i)
+            for (var i = 0; i < length; ++i)
             {
                 var quest = SerializedQuest.FromArray(Reader.ReadBytes(Reader.ReadInt32()));
-                if(MissionPool.Exists(quest.Name) || quest.IsMetadata)
+                if (MissionPool.Exists(quest.Name) || quest.IsMetadata)
                     quests.Add(quest);
                 else
                     Log.WriteLine($"Found non-existent quest design '{quest.Name}', removing...");
             }
+
             return quests.ToArray();
         }
-        
+
         private static string[] LoadRecipes(BinaryReader Reader)
         {
             var recipes = new List<string>();
             var length = Reader.ReadInt32();
-            for(var i = 0; i < length; ++i)
+            for (var i = 0; i < length; ++i)
             {
                 var name = Reader.ReadString();
-                if(ItemPool.Exists(name))
+                if (ItemPool.Exists(name))
                     recipes.Add(name);
                 else
                     Log.WriteLine($"Found non-existent recipe '{name},' removing...");
@@ -237,7 +274,7 @@ namespace Hedra.Engine.Management
 
             return recipes.ToArray();
         }
-        
+
         private static KeyValuePair<int, Item>[] LoadItems(BinaryReader Reader)
         {
             var items = new Dictionary<int, Item>();
@@ -247,64 +284,19 @@ namespace Hedra.Engine.Management
                 var index = Reader.ReadInt32();
                 var item = Item.FromArray(Reader.ReadBytes(Reader.ReadInt32()));
                 if (item != null)
-                {
                     items.Add(index, item);
-                }
                 else
-                {
-                    Log.WriteLine($"Found non-existent item, removing...");
-                }
+                    Log.WriteLine("Found non-existent item, removing...");
             }
+
             return items.ToArray();
         }
-        
-        
+
+
         public static void DeleteCharacter(PlayerInformation Information)
         {
             File.Delete($"{CharactersFolder}{Information.Name}.db");
             File.Delete($"{CharactersFolder}{Information.Name}.db.bak");
         }
-
-        public static PlayerInformation[] PlayerFiles
-        {
-            get
-            {
-                var filesList = new List<PlayerInformation>();
-                string[] files = Directory.GetFiles(CharactersFolder);
-
-                for(var i = 0; i < files.Length; i++)
-                {
-                    if(files[i].EndsWith(".bak")) 
-                        continue;
-                    try
-                    {
-                        var information = LoadPlayer(files[i]);
-                        if (information == null)
-                        {
-                            File.Delete(files[i]);
-                            if (File.Exists(files[i] + ".bak"))
-                            {
-                                File.Delete(files[i] + ".bak");
-                            }
-                            continue;
-                        }
-                        filesList.Add( information );
-                    }
-                    catch (Exception e)
-                    {
-                        if(e is UnauthorizedAccessException)
-                        {
-                            Log.WriteLine("Unauthorized access in file: " + Path.GetFileName(files[i]));
-                            continue;
-                        }
-                        Log.WriteLine(e.ToString());
-                    }
-                }
-                return filesList.ToArray();
-            }
-        }
-        
-        
-        public static int CharacterCount => PlayerFiles.Length;
     }
 }
